requirements.txt
Flask
pyusb
qrcode[pil]
pyinstaller
paramiko
pywifi
scapy
requests
beautifulsoup4
libvirt-python
opencv-python
discord.py